app/requirements.txt
numpy==1.26.4
opencv-python==4.10.0.84
PySide6==6.6.2
pandas==2.2.3
tensorflow==2.19.0
tf-keras
deepface==0.0.79
mtcnn
retina-face
gdown
requests
tqdm
mss
Pillow==10.4.0
